+++ table/block_based/filter_block_reader_common.cc
@@ -8,10 +8,17 @@
 
 #include "block_cache.h"
 #include "monitoring/perf_context_imp.h"
+#include "rocksdb/table_pinning_policy.h"
 #include "table/block_based/block_based_table_reader.h"
 #include "table/block_based/parsed_full_filter_block.h"
 
 namespace ROCKSDB_NAMESPACE {
+template <typename TBlocklike>
+FilterBlockReaderCommon<TBlocklike>::~FilterBlockReaderCommon() {
+  if (pinned_) {
+    table_->UnPinData(std::move(pinned_));
+  }
+}
 
 template <typename TBlocklike>
 Status FilterBlockReaderCommon<TBlocklike>::ReadFilterBlock(