+++ table/block_based/filter_block_reader_common.h
@@ -15,6 +15,7 @@ namespace ROCKSDB_NAMESPACE {
 
 class BlockBasedTable;
 class FilePrefetchBuffer;
+struct PinnedEntry;
 
 // Encapsulates common functionality for the various filter block reader
 // implementations. Provides access to the filter block regardless of whether
@@ -24,8 +25,11 @@ template <typename TBlocklike>
 class FilterBlockReaderCommon : public FilterBlockReader {
  public:
   FilterBlockReaderCommon(const BlockBasedTable* t,
-                          CachableEntry<TBlocklike>&& filter_block)
-      : table_(t), filter_block_(std::move(filter_block)) {
+                          CachableEntry<TBlocklike>&& filter_block,
+                          std::unique_ptr<PinnedEntry>&& pinned)
+      : table_(t),
+        filter_block_(std::move(filter_block)),
+        pinned_(std::move(pinned)) {
     assert(table_);
     const SliceTransform* const prefix_extractor = table_prefix_extractor();
     if (prefix_extractor) {
@@ -33,7 +37,7 @@ class FilterBlockReaderCommon : public FilterBlockReader {
           prefix_extractor->FullLengthEnabled(&prefix_extractor_full_length_);
     }
   }
-
+  ~FilterBlockReaderCommon() override;
   bool RangeMayExist(const Slice* iterate_upper_bound, const Slice& user_key,
                      const SliceTransform* prefix_extractor,
                      const Comparator* comparator,
@@ -69,6 +73,7 @@ class FilterBlockReaderCommon : public FilterBlockReader {
  private:
   const BlockBasedTable* table_;
   CachableEntry<TBlocklike> filter_block_;
+  std::unique_ptr<PinnedEntry> pinned_;
   size_t prefix_extractor_full_length_ = 0;
   bool full_length_enabled_ = false;
 };