+++ table/block_based/hash_index_reader.cc
@@ -8,12 +8,14 @@
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 #include "table/block_based/hash_index_reader.h"
 
+#include "rocksdb/table_pinning_policy.h"
 #include "table/block_fetcher.h"
 #include "table/meta_blocks.h"
 
 namespace ROCKSDB_NAMESPACE {
 Status HashIndexReader::Create(const BlockBasedTable* table,
                                const ReadOptions& ro,
+                               const TablePinningOptions& tpo,
                                FilePrefetchBuffer* prefetch_buffer,
                                InternalIterator* meta_index_iter,
                                bool use_cache, bool prefetch, bool pin,
@@ -26,6 +28,7 @@ Status HashIndexReader::Create(const BlockBasedTable* table,
   const BlockBasedTable::Rep* rep = table->get_rep();
   assert(rep != nullptr);
 
+  std::unique_ptr<PinnedEntry> pinned;
   CachableEntry<Block> index_block;
   if (prefetch || !use_cache) {
     const Status s =
@@ -35,7 +38,11 @@ Status HashIndexReader::Create(const BlockBasedTable* table,
       return s;
     }
 
-    if (use_cache && !pin) {
+    if (pin) {
+      table->PinData(tpo, TablePinningPolicy::kIndex,
+                     index_block.GetValue()->ApproximateMemoryUsage(), &pinned);
+    }
+    if (use_cache && !pinned) {
       index_block.Reset();
     }
   }
@@ -44,7 +51,8 @@ Status HashIndexReader::Create(const BlockBasedTable* table,
   // hard error. We can still fall back to the original binary search index.
   // So, Create will succeed regardless, from this point on.
 
-  index_reader->reset(new HashIndexReader(table, std::move(index_block)));
+  index_reader->reset(
+      new HashIndexReader(table, std::move(index_block), std::move(pinned)));
 
   // Get prefixes block
   BlockHandle prefixes_handle;