+++ table/block_based/partitioned_index_reader.cc
@@ -10,14 +10,16 @@
 
 #include "block_cache.h"
 #include "file/random_access_file_reader.h"
+#include "rocksdb/table_pinning_policy.h"
 #include "table/block_based/block_based_table_reader.h"
 #include "table/block_based/partitioned_index_iterator.h"
 
 namespace ROCKSDB_NAMESPACE {
 Status PartitionIndexReader::Create(
     const BlockBasedTable* table, const ReadOptions& ro,
-    FilePrefetchBuffer* prefetch_buffer, bool use_cache, bool prefetch,
-    bool pin, BlockCacheLookupContext* lookup_context,
+    const TablePinningOptions& tpo, FilePrefetchBuffer* prefetch_buffer,
+    bool use_cache, bool prefetch, bool pin,
+    BlockCacheLookupContext* lookup_context,
     std::unique_ptr<IndexReader>* index_reader) {
   assert(table != nullptr);
   assert(table->get_rep());
@@ -25,6 +27,7 @@ Status PartitionIndexReader::Create(
   assert(index_reader != nullptr);
 
   CachableEntry<Block> index_block;
+  std::unique_ptr<PinnedEntry> pinned;
   if (prefetch || !use_cache) {
     const Status s =
         ReadIndexBlock(table, prefetch_buffer, ro, use_cache,
@@ -33,12 +36,18 @@ Status PartitionIndexReader::Create(
       return s;
     }
 
-    if (use_cache && !pin) {
+    if (pin) {
+      pin = table->PinData(tpo, TablePinningPolicy::kTopLevel,
+                           index_block.GetValue()->ApproximateMemoryUsage(),
+                           &pinned);
+    }
+    if (use_cache && !pinned) {
       index_block.Reset();
     }
   }
 
-  index_reader->reset(new PartitionIndexReader(table, std::move(index_block)));
+  index_reader->reset(new PartitionIndexReader(table, std::move(index_block),
+                                               std::move(pinned)));
 
   return Status::OK();
 }