+++ table/sst_file_dumper.cc
@@ -172,7 +172,7 @@ Status SstFileDumper::NewTableReader(
   auto t_opt = TableReaderOptions(
       ioptions_, moptions_.prefix_extractor, soptions_, internal_comparator_,
       0 /* block_protection_bytes_per_key */, false /* skip_filters */,
-      false /* immortal */, true /* force_direct_prefetch */, -1 /* level */,
+      false /* immortal */, true /* force_direct_prefetch */, -1 /* level */, false /* bottommost */,
       nullptr /* block_cache_tracer */, 0 /* max_file_size_for_l0_meta_pin */,
       "" /* cur_db_session_id */, 0 /* cur_file_num */, {} /* unique_id */,
       0 /* largest_seqno */, 0 /* tail_size */,